backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
duckdb
pandas
google-generativeai
psycopg2-binary
SqlAlchemy
openpyxl
PyMySQL
boto3
s3fs